// app/requests/_personal/_moderation/getProject/getProjectSuccess.type.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GetProjectSuccess {
  data: {
    id: number;
    name: string;
    brand: {
      id: number;
      name: string;
      logo: string | null;
      description: string;
    }[];
  }[];
}